# generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST Library API for Ximango Resourses",
    "version" : "version 1.0",
    "title" : "REST API Documentation for WPLEX REST Library",
    "termsOfService" : "http://www.wplex.com.br/terms",
    "contact" : {
      "name" : "Ryan Padilha",
      "url" : "http",
      "email" : "ryan.padilha@wplex.com.br"
    },
    "license" : {
      "name" : "WPLEX License Terms",
      "url" : "http://www.wplex.com.br/license"
    }
  },
  "host" : "localhost:8080",
  "tags" : [ {
    "name" : "company"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/api/v1/company/" : {
      "get" : {
        "tags" : [ "company" ],
        "summary" : "Retrieves a list of companies",
        "description" : "",
        "operationId" : "list",
        "responses" : {
          "200" : {
            "description" : "Retrieves a list of companies",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Company"
              }
            }
          },
          "204" : {
            "description" : "No content retrieve"
          }
        }
      },
      "post" : {
        "tags" : [ "company" ],
        "summary" : "Create a company entity",
        "description" : "",
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "company",
          "description" : "Company json stream entity",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Company"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Create a new company entity",
            "schema" : {
              "$ref" : "#/definitions/Company"
            }
          },
          "304" : {
            "description" : "Return a entity not modified",
            "schema" : {
              "$ref" : "#/definitions/Company"
            }
          }
        }
      }
    },
    "/api/v1/company/{id}" : {
      "get" : {
        "tags" : [ "company" ],
        "summary" : "Get a company by id",
        "description" : "",
        "operationId" : "get",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Company Id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Retrieve a company searched by id",
            "schema" : {
              "$ref" : "#/definitions/Company"
            }
          },
          "404" : {
            "description" : "Not found retrieve searched by id"
          }
        }
      },
      "put" : {
        "tags" : [ "company" ],
        "summary" : "Update a company entity",
        "description" : "",
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Company Id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Company json stream entity",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Company"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Retrieve an update a company entity",
            "schema" : {
              "$ref" : "#/definitions/Company"
            }
          },
          "404" : {
            "description" : "Not found retrieve if no update was process"
          }
        }
      },
      "delete" : {
        "tags" : [ "company" ],
        "summary" : "Delete a company entity",
        "description" : "",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Company Id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "204" : {
            "description" : "No contente retive, company deleted",
            "schema" : {
              "$ref" : "#/definitions/Company"
            }
          },
          "404" : {
            "description" : "Not found retrieve if no delete was process"
          }
        }
      }
    }
  },
  "definitions" : {
    "Company" : {
      "type" : "object",
      "required" : [ "created", "initials", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "name" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 100
        },
        "initials" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 5
        }
      }
    }
  }
}